requirements.txt
alembic==1.0.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.13
chardet==3.0.4
idna==2.7
Mako==1.0.7
MarkupSafe==1.0
pyaml==17.12.1
pymongo==3.7.1
python-dateutil==2.7.3
python-editor==1.0.3
requests==2.20.0
scraperwiki==0.5.1
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
